requirements.txt
black
boto3
duckdb
isort
lightgbm
matplotlib
numpy
pandas
python-dotenv
scikit-learn
scipy
seaborn
shap
streamlit
tqdm
PyYAML